Re-Translates the Lewd Module - Everything up to lewd_items
The old translation made a lot of these stunted; and we don't exactly have access to the original Russian text anymore. As such; I've opted to outright re-write the majority of the text alongside fixing up code comments. Additionally removes deprecated code for the long-dead nymphomaniac quirk AND the coom item.

Generally speaking, less assumptions are made in text relating to character states and the lack of 'equipment'.
Gets rid of the domina cap text when worn / dropped. It's out of place.
Implements a NEW funny easter egg for hypnogoggles. April Fool's only; and won't interrupt anyone actually wanting to use them
Blindfolds are more sane on when to display their text